Add brief documentation on coding conventions
The first level of contribution guidelines have been documented in the
CONTRIBUTING.md file. One possible set of coding conventions and
development workflow have been introduced. At this stage, it is
important to note that these guidelines are just suggestions and are
open to discussion for improvement.

This commit has been squashed from the `issue/2` branch, and is also
being used to demonstrate the squash and commit workflow through an
intermediary staging branch.