Start implementing configure script
The configure script is being developed to ensure that the
packages required to build RefPerSys are available. As of now, we
are assuming a Debian 11 host machine. The configure script also
supports the -h, -l and -v flags to display the help, license and
version information.

Additionally, we are specifying gcc-10 and
g++-10 in our generated local ~/.refpersys.mk file (since these
are available by default on Debian 11).

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>