Add compatiblity macros
The <sys/cdefs.h> head file provides several useful optimisiation
macros. These macros are expected to be present, but if not, we
take care to define them since they will be used quite often. In
many cases C code on FreeBSD and OpenBSD refer to the non-standard
explicit_bzero() and strlcpy() functions, and we're mapping them to
standard memset() and strncpy() in case they are unavailable.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>