Declare log interface for *BSD port
Logging is performed in the mainline branch through utility
functions such as rps_fatal_stop_at(). The log interface on the
*BSD port adapts this idea and extends it to provide an interface
to log messages to either the syslog or a specific file in addition
to printing onto a virtual or pseudo TTY. The implementation of the
interface will be completed in a subsequent commit.

Another consideration is the use of libbacktrace in the mainline
source code. Although libbacktrace mentions that it should support
the *BSDs, unfortunately it has not yet found itself into the *BSD
ports. Possible workarounds that could be explored include:
  - backtrace(3), which is supported on Linux and *BSD
  - the non-GNU libunwind, but its support for clang is not known
  - LLVM libunwind
  - C++23 std::stacktrace when it becomes better available

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>