Ensure __thread storage class is available
Thread-local storage would be required for the a few variables,
especially in managers such as the log manager. Although __thread
is most likely available, we are nonetheless taking care to ensure
that it exists. If it does not exist, we fall back to the standard
C11 thread_local storage class.

Signed-off-by: Abhishek Chakravarti <abhishek@taranjali.org>